Rename project to scylla-api-client
Following bhalevy's request, the name of the project is renamed to
scylla-api-client.
